Add Linux build support (#107)
This PR is a quick detour from my work moving the compilation artifacts over to sclang-accessible structures. The code needed only minor changes to compile with GCC, and I also had to add a few OS-specific functions to find the running binary path and map executable heap pages. Now that Hadron compiles on two operating systems, it's a priority to get the continuous integration setup (#30) to ensure that both builds remain healthy.

+++ src/hadron/Slot.hpp
@@ -44,7 +44,7 @@ public:
     static inline Slot makeFloat(double d) { return Slot(d); }
     static inline Slot makeNil() { return Slot(kObjectPointerTag); }
     static inline Slot makeInt32(int32_t i) { return Slot(kInt32Tag | (static_cast<uint64_t>(i) & (~kTagMask))); }
-    static inline Slot makeBool(bool b) { return Slot(kBooleanTag | (b ? 1ull : 0ull)); }
+    static inline Slot makeBool(bool b) { return Slot(static_cast<uint64_t>(kBooleanTag | (b ? 1ull : 0ull))); }
     static inline Slot makePointer(library::Schema* p) {
         return Slot(kObjectPointerTag | reinterpret_cast<uint64_t>(p));
     }
